iOS: Emphasize headings based on structure
Basically, the first level that has multiple headings gets bolded (and any parent levels it).
In practice, this means the most significant headings in the page get bolded; their children remain normal; and if there's something like a single <h1> at the top, it also gets bolded, for good looks.