iOS: Don't immediately reveal tapped page heading
The tap causes the table of contents to scroll, which in turn causes different headings to be revealed in the page, one after another, until the tapped heading is reached.
This means that the target heading would be revealed, but then very quickly replaced by another heading—it just looked glitchy.