Shared: Fix scroll sliding when zoomed in in Safari
In Safari 17.2.1, when the page is zoomed, getBoundingClientRect() starts reporting fractional values. (which also happen to be incorrect, but that's another bug)

When repeatedly scrolling to the same heading (e.g. when click-and-dragging in the heading list), this results in the computed scrollby amount oscillating between very small values, which continuously scrolls the page slightly. This commit prevents that.